refactor(format): extract Markdown renderer into tokmd-analysis-format-md microcrate (#998)
Extract ~850 lines of Markdown formatting logic from tokmd-analysis-format
into a dedicated tokmd-analysis-format-md microcrate.

Changes:
- Create crates/tokmd-analysis-format-md/ with render_md() entry point
- tokmd-analysis-format now delegates to microcrate via pub re-export
- Add microcrate to workspace members and dependencies

Behavioral improvements:
- Empty doc_density_by_lang section now correctly omitted (was rendering
  empty table headers)
- Minor spacing fix in polyglot output (trailing space removed)

Verification:
- All 400+ tests pass across both crates
- Zero unwrap/expect/unsafe in new microcrate
- Full workspace builds clean
- 2 unused import warnings fixed

Refs #998
